+++ kwimage/structs/_generic.py
@@ -126,12 +126,12 @@ class ObjectList(Spatial):
         newdata = [None if item is None else func(item) for item in self.data]
         return self.__class__(newdata, self.meta)
 
-    def to_coco(self):
+    def to_coco(self, style='orig'):
         for item in self.data:
             if item is None:
                 yield None
             else:
-                yield item.to_coco()
+                yield item.to_coco(style=style)
 
     def compress(self, flags, axis=0):
         assert axis == 0
@@ -164,8 +164,35 @@ class ObjectList(Spatial):
         return self.apply(lambda item: item.numpy())
 
     @classmethod
-    def concatenate(cls, data):
-        raise NotImplementedError
+    def concatenate(cls, items, axis=0):
+        """
+        Args:
+            items (Sequence[ObjectList]): multiple object lists of the same type
+            axis (int | None): unused, always implied to be axis 0
+
+        Returns:
+            ObjectList: combined object list
+
+        Example:
+            >>> import kwimage
+            >>> cls = kwimage.MaskList
+            >>> sub_cls = kwimage.Mask
+            >>> item1 = cls([sub_cls.random(), sub_cls.random()])
+            >>> item2 = cls([sub_cls.random()])
+            >>> items = [item1, item2]
+            >>> new = cls.concatenate(items)
+            >>> assert len(new) == 3
+        """
+        if len(items) == 0:
+            new = cls([])
+        else:
+            newdata = []
+            for item in items:
+                newdata.extend(item.data)
+
+        newmeta = items[0].meta
+        new = cls(newdata, newmeta)
+        return new
 
     def is_tensor(cls):
         raise NotImplementedError
@@ -207,3 +234,45 @@ def _safe_compress(data, flags, axis):
         return data.compress(flags, axis=axis)
     except (TypeError, AttributeError):
         return kwarray.ArrayAPI.compress(data, flags, axis=axis)
+
+
+def _issubclass2(child, parent):
+    """
+    Uses string comparisons to avoid ipython reload errors.
+    Much less robust though.
+    """
+    # String comparison
+    if child.__name__ == parent.__name__:
+        if child.__module__ == parent.__module__:
+            return True
+    # Recurse through classes of obj
+    return any(_issubclass2(base, parent) for base in child.__bases__)
+
+
+def _isinstance2(obj, cls):
+    """
+    Uses string comparisons to avoid ipython reload errors.
+    Much less robust though.
+
+    Example:
+        import kwimage
+        from kwimage.structs import _generic
+        cls = kwimage.structs._generic.ObjectList
+        obj = kwimage.MaskList([])
+        _generic._isinstance2(obj, cls)
+
+        _generic._isinstance2(kwimage.MaskList([]), _generic.ObjectList)
+
+        dets = kwimage.Detections(
+            boxes=kwimage.Boxes.random(3).numpy(),
+            class_idxs=[0, 1, 1],
+            segmentations=kwimage.MaskList([None] * 3)
+        )
+    """
+    if isinstance(obj, cls):
+        return True
+    try:
+        return _issubclass2(obj.__class__, cls)
+    except Exception:
+        return False
+    return False