+++ kwimage/structs/detections.py
@@ -110,6 +110,10 @@ class _DetDrawMixin:
             >>> kwplot.autompl()
             >>> kwplot.imshow(image2)
             >>> kwplot.show_if_requested()
+
+        Ignore:
+            import xdev
+            globals().update(xdev.get_func_kwargs(kwimage.Detections.draw_on))
         """
         labels = self._make_labels(labels)
         alpha = self._make_alpha(alpha)
@@ -487,9 +491,9 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
                 other = []
                 objlist = []
                 for k, v in data.items():
-                    if isinstance(v, _generic.ObjectList):
+                    if _generic._isinstance2(v, _generic.ObjectList):
                         objlist.append(v)
-                    elif isinstance(v, _boxes.Boxes):
+                    elif _generic._isinstance2(v, _boxes.Boxes):
                         if v.is_numpy():
                             ndarrays.append(k)
                         else:
@@ -697,7 +701,7 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
                 dets.meta['kp_classes'] = kp_classes
         return dets
 
-    def to_coco(self, cname_to_cat=None):
+    def to_coco(self, cname_to_cat=None, style='orig'):
         """
         Converts this set of detections into coco-like annotation dictionaries.
 
@@ -716,6 +720,9 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
         Args:
             cname_to_cat: currently ignored.
 
+            style (str): either orig (for the original coco format) or new
+                for the more general ndsampler-style coco format.
+
         Yields:
             dict: coco-like annotation structures
 
@@ -729,7 +736,7 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
         import kwarray
         to_collate = {}
         if 'boxes' in self.data:
-            to_collate['bbox'] = list(self.data['boxes'].to_coco())
+            to_collate['bbox'] = list(self.data['boxes'].to_coco(style=style))
 
         if 'class_idxs' in self.data:
             if 'classes' in self.meta:
@@ -743,10 +750,12 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
                     self.data['class_idxs'])
 
         if 'keypoints' in self.data:
-            to_collate['keypoints'] = list(self.data['keypoints'].to_coco())
+            to_collate['keypoints'] = list(self.data['keypoints'].to_coco(
+                style=style))
 
         if 'segmentations' in self.data:
-            to_collate['segmentation'] = list(self.data['segmentations'].to_coco())
+            to_collate['segmentation'] = list(self.data['segmentations'].to_coco(
+                style=style))
 
         if 'scores' in self.data:
             to_collate['score'] = kwarray.ArrayAPI.tolist(self.data['scores'])
@@ -880,6 +889,12 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
             >>> dets = [self, other]
             >>> new = Detections.concatenate(dets)
             >>> assert new.num_boxes() == 5
+
+            >>> self = Detections.random(2, segmentations=True)
+            >>> other = Detections.random(3, segmentations=True)
+            >>> dets = [self, other]
+            >>> new = Detections.concatenate(dets)
+            >>> assert new.num_boxes() == 5
         """
         if len(dets) == 0:
             raise ValueError('need at least one detection to concatenate')
@@ -902,7 +917,6 @@ class Detections(ub.NiceRepr, _DetAlgoMixin, _DetDrawMixin):
                     msg = ('Error when trying to concat {}'.format(key))
                     print(msg)
                     raise
-                    raise Exception(msg)
 
         newmeta = dets[0].meta
         new = cls(newdata, newmeta)