+++ CHANGELOG.md
@@ -3,17 +3,20 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
+## Version 0.5.5 -
+
 ## Version 0.5.4 - Released 2019-Dec-18
 
 ### Added
 * Tentative `Color._forimage` method
 
 ### Changed
-* Simpler demod data file names.
+* Simpler demo data file names.
 * The kwimage.struct `draw_on` methods now operate in-place (more often) when possible.
 
 ### Fixed
 * Fixed color in kwimage.struct `draw_on` methods.
+* Detections concatenate now works when segmentations are populated.
 
 
 ## Version 0.5.3 - Released 2019-Dec-17